room enter and disconnect made
when enter to room, icon appear
when disconnect, icon disappear
controller modified, join_room deleted
game controller room_connect modified

+++ app/controllers/game_controller.rb
@@ -9,8 +9,17 @@ class GameController < WebsocketRails::BaseController
 
   def room_connect
     if Room.find_by_code(message) != nil
+      puts message
+      room_id = Room.select(:id).where(code: message).limit(1)[0].id
+      unless Player.where(user_id: current_user.id, room_id: room_id).exists?
+         player = Player.new
+         player.user_id = current_user.id
+         player.room_id = room_id
+         player.username = current_user.username
+         player.save
+         WebsocketRails[("room_"+message).to_sym].trigger(:player_enter, {player_info: player})
+      end
       send_message :room_connect_status, {message: 'room connection success', status:true}, :namespace => 'game'
-
     else
       send_message :room_connect_status, {message: 'There is no room', status:false}, :namespace => 'game'
     end
@@ -22,11 +31,14 @@ class GameController < WebsocketRails::BaseController
   end
 
   def room_disconnect
-    Player.where(user_id: current_user.id).destroy_all
-    WebsocketRails[("room_"+message["room_code"]).to_sym].trigger(:player_disconnect, {id: current_user.id})
-    room = Room.where(code: message["room_code"]).limit(1)[0]
-    if room.players.length == 0
-      room.destroy
+    player = Player.where(user_id: current_user.id).limit(1)[0]
+    room   = Room.where(code: message["room_code"]).limit(1)[0]
+    if player.destroy
+      WebsocketRails[("room_"+message["room_code"]).to_sym].trigger(:player_disconnect, {player_info: player})
+
+      if room.players.length == 0
+        room.destroy
+      end
     end
   end
